Search the current master when starting
When a container is started, the master is sometimes no longer the
rancher's primary ip, due to a failover that switched the master. A
slave would try to connect on another slave. The startup script searches
for the current master, and if it can't find any (first start), then, it
chooses the primary ip.